feat: Refactor user signup and signin routes
The code changes refactor the user signup and signin routes in the `user.js` file. The unnecessary line of code in the signup route has been removed. Additionally, the signin route has been modified to use the `findOne` method with a single line of code. This improves the efficiency and readability of the code.

Note: Please remove any meta information such as issue references, tags, or author names before using the commit message.