chore: Add build script for server
This commit adds a build script for the server in the package.json file. The build script uses esbuild to compile the TypeScript code in the src/index.ts file and outputs the bundled JavaScript code to the dist directory. This will improve the development workflow by providing a pre-built version of the server code that can be easily run.