chore: Update file paths in package.json
This commit updates the file paths in the package.json files of the db and store packages. The paths for the "./client" and "./balance" exports have been modified to "./src/index.ts" and "./src/hooks/useBalance.ts" respectively. This change ensures that the correct files are being exported and used in the project.